Fix docker metrics discovery by converting port to int64

+++ libbeat/autodiscover/providers/docker/docker.go
@@ -150,7 +150,7 @@ func (d *Provider) emitContainer(event bus.Event, flag string) {
 		event := bus.Event{
 			flag:     true,
 			"host":   host,
-			"port":   port.PrivatePort,
+			"port":   int64(port.PrivatePort),
 			"docker": meta,
 			"meta": common.MapStr{
 				"docker": meta,
